+++ src/shared/ui/Calendar/index.ts
@@ -1,3 +1,4 @@
 export { default as Calendar } from "./Calendar";
-export type { AttendanceStatus, DateData } from "./Calendar";
+export type { AttendanceStatus } from "./Calendar";
+export type { CalendarAttendanceStatus, DateData } from "./types";
 export { default as CalendarSkeleton } from "./CalendarSkeleton";
